Extract Dotty LSP integration into separate package.
We're going to stick with the "add the Dotty depends to the package depends"
approach for now because it's simple and fairly robust, but I don't want to
permanently saddle the scala-mode package with the rather copious Dotty
depends.

This is more like how things work in VS Code anyway, you install a separate
package that contains langserver bindings for some langserver you want to use.